Wait for scripts to load before rendering
Also moves plugin code into plugins.tsx
